Ask for permissions during options saving
Seems like a more reliable way
