+++ src/js/background.js
@@ -1,7 +1,7 @@
 /* @flow */
 // TODO do I really need to annotate all files with @flow??
 
-import {COMMAND_CAPTURE_SIMPLE, METHOD_CAPTURE_WITH_EXTRAS, showNotification, ensurePermissions} from './common';
+import {COMMAND_CAPTURE_SIMPLE, METHOD_CAPTURE_WITH_EXTRAS, showNotification} from './common';
 import {get_options} from './options';
 
 type Params = {
@@ -108,7 +108,7 @@ function capture(comment: ?string = null, tag_str: ?string = null) {
 
 chrome.commands.onCommand.addListener(command => {
     if (command === COMMAND_CAPTURE_SIMPLE) {
-        ensurePermissions(() => capture(null, null));
+        capture(null, null);
     }
 });
